third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-move-breakpoints-expected.txt

Verify that breakpoints are moved appropriately


Running: addFileSystem

Running: setBreakpointInFileSystemUISourceCode
    file:///var/www/inspector/persistence/resources/foo.js:0

Running: addFileMapping
    http://127.0.0.1:8000/inspector/persistence/resources/foo.js:0

Running: removeFileMapping
    http://127.0.0.1:8000/inspector/persistence/resources/foo.js:0
    file:///var/www/inspector/persistence/resources/foo.js:0
